refactor: split event page client boundaries
