+++ src/app/[locale]/(platform)/event/[slug]/_components/EventRules.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 import type { Event } from '@/types'
 import { LinkIcon } from 'lucide-react'
 import { useExtracted, useLocale } from 'next-intl'
