+++ src/i18n/messages/de.json
@@ -112,6 +112,7 @@
   "E2h9bS": "gekauft",
   "JRKiwS": "verkauft",
   "dTR5nN": "zu",
+  "Gs9Kfp": "Nach oben",
   "BWpuKl": "Aktualisierung",
   "recCg9": "Aktualisierungen",
   "P48sLl": "Ereignisänderungen",
@@ -178,8 +179,6 @@
   "6mRa9V": "Am meisten gemocht",
   "Riqzb0": "Inhaber",
   "fkEYTF": "Noch keine Kommentare. Sei der/die Erste!",
-  "KQsZBW": "Ähnlich",
-  "Gs9Kfp": "Nach oben",
   "oUWADl": "Nein",
   "a5msuh": "Ja",
   "QA1wrZ": "Häufig gestellte Fragen",
@@ -323,6 +322,7 @@
   "beB5uE": "Anteile",
   "k/3gzq": "Ähnliche Ereignisse konnten nicht geladen werden.",
   "AqlPcP": "Für dieses Tag gibt es noch keine verwandten Ereignisse.",
+  "KQsZBW": "Ähnlich",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "Resolver",
   "kYQE3Q": "Abwicklung vorschlagen",
