+++ src/i18n/messages/en.json
@@ -112,6 +112,7 @@
   "E2h9bS": "bought",
   "JRKiwS": "sold",
   "dTR5nN": "at",
+  "Gs9Kfp": "Back to top",
   "BWpuKl": "Update",
   "recCg9": "Updates",
   "P48sLl": "Event changes",
@@ -178,8 +179,6 @@
   "6mRa9V": "Most liked",
   "Riqzb0": "Holders",
   "fkEYTF": "No comments yet. Be the first to comment!",
-  "KQsZBW": "Related",
-  "Gs9Kfp": "Back to top",
   "oUWADl": "No",
   "a5msuh": "Yes",
   "QA1wrZ": "Frequently Asked Questions",
@@ -323,6 +322,7 @@
   "beB5uE": "shares",
   "k/3gzq": "Failed to fetch related events.",
   "AqlPcP": "No related events for this tag yet.",
+  "KQsZBW": "Related",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "Resolver",
   "kYQE3Q": "Propose resolution",
