+++ src/i18n/messages/es.json
@@ -112,6 +112,7 @@
   "E2h9bS": "compró",
   "JRKiwS": "vendió",
   "dTR5nN": "a",
+  "Gs9Kfp": "Volver arriba",
   "BWpuKl": "Actualización",
   "recCg9": "Actualizaciones",
   "P48sLl": "Cambios del evento",
@@ -178,8 +179,6 @@
   "6mRa9V": "Más gustados",
   "Riqzb0": "Titulares",
   "fkEYTF": "Aún no hay comentarios. ¡Sé el primero en comentar!",
-  "KQsZBW": "Relacionado",
-  "Gs9Kfp": "Volver arriba",
   "oUWADl": "No",
   "a5msuh": "Sí",
   "QA1wrZ": "Preguntas frecuentes",
@@ -323,6 +322,7 @@
   "beB5uE": "acciones",
   "k/3gzq": "No se pudieron cargar los eventos relacionados.",
   "AqlPcP": "Aún no hay eventos relacionados para esta etiqueta.",
+  "KQsZBW": "Relacionado",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "Resolutor",
   "kYQE3Q": "Proponer resolución",
