+++ src/i18n/messages/fr.json
@@ -112,6 +112,7 @@
   "E2h9bS": "a acheté",
   "JRKiwS": "a vendu",
   "dTR5nN": "à",
+  "Gs9Kfp": "Retour en haut",
   "BWpuKl": "Mise à jour",
   "recCg9": "Mises à jour",
   "P48sLl": "Modifications de l’événement",
@@ -178,8 +179,6 @@
   "6mRa9V": "Les plus aimés",
   "Riqzb0": "Détenteurs",
   "fkEYTF": "Aucun commentaire pour l’instant. Soyez le premier à commenter !",
-  "KQsZBW": "Associé",
-  "Gs9Kfp": "Retour en haut",
   "oUWADl": "Non",
   "a5msuh": "Oui",
   "QA1wrZ": "Questions fréquentes",
@@ -323,6 +322,7 @@
   "beB5uE": "parts",
   "k/3gzq": "Impossible de charger les événements associés.",
   "AqlPcP": "Aucun événement associé pour ce tag pour le moment.",
+  "KQsZBW": "Associé",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "Résolveur",
   "kYQE3Q": "Proposer une résolution",
