+++ src/i18n/messages/pt.json
@@ -112,6 +112,7 @@
   "E2h9bS": "comprou",
   "JRKiwS": "vendeu",
   "dTR5nN": "a",
+  "Gs9Kfp": "Voltar ao topo",
   "BWpuKl": "Atualização",
   "recCg9": "Atualizações",
   "P48sLl": "Alterações do evento",
@@ -178,8 +179,6 @@
   "6mRa9V": "Mais curtidos",
   "Riqzb0": "Detentores",
   "fkEYTF": "Ainda não há comentários. Seja o primeiro a comentar!",
-  "KQsZBW": "Relacionado",
-  "Gs9Kfp": "Voltar ao topo",
   "oUWADl": "Não",
   "a5msuh": "Sim",
   "QA1wrZ": "Perguntas frequentes",
@@ -323,6 +322,7 @@
   "beB5uE": "posições",
   "k/3gzq": "Não foi possível carregar eventos relacionados.",
   "AqlPcP": "Ainda não há eventos relacionados para esta tag.",
+  "KQsZBW": "Relacionado",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "Resolvedor",
   "kYQE3Q": "Propor resolução",
