+++ src/i18n/messages/zh.json
@@ -112,6 +112,7 @@
   "E2h9bS": "买入",
   "JRKiwS": "卖出",
   "dTR5nN": "于",
+  "Gs9Kfp": "返回顶部",
   "BWpuKl": "更新",
   "recCg9": "更新",
   "P48sLl": "事件变更",
@@ -178,8 +179,6 @@
   "6mRa9V": "最受喜欢",
   "Riqzb0": "持有人",
   "fkEYTF": "还没有评论，来抢沙发！",
-  "KQsZBW": "相关",
-  "Gs9Kfp": "返回顶部",
   "oUWADl": "否",
   "a5msuh": "是",
   "QA1wrZ": "常见问题",
@@ -323,6 +322,7 @@
   "beB5uE": "份额",
   "k/3gzq": "无法获取相关事件。",
   "AqlPcP": "该标签暂无相关事件。",
+  "KQsZBW": "相关",
   "SG8ntl": "0x0000...0000",
   "hNq/TA": "裁决器",
   "kYQE3Q": "提出结算",
